backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9        # multipart/form-data (file uploads)

# ── Database ─────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0                # async PostgreSQL driver
pgvector>=0.3.0                # pgvector SQLAlchemy / psycopg integration

# ── Validation / settings ─────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0
email-validator>=2.1.0         # required for pydantic EmailStr

# ── Auth / security ───────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
supabase>=2.4.0

# ── HTTP ─────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── AI / RAG ─────────────────────────────────────────────────────────
anthropic>=0.28.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
langchain-community>=0.2.0

# ── Cache / queue ─────────────────────────────────────────────────────
redis>=5.0.4
arq>=0.25.0

# ── Storage ───────────────────────────────────────────────────────────
boto3>=1.34.0                  # R2 via S3-compatible API
